requirements.txt
Django==1.7.3
PyJWT==0.4.1
django-angular==0.7.10
django-braces==1.4.0
django-crispy-forms==1.4.0
django-disqus==0.4.3
django-extensions==1.4.9
django-pagination==1.0.7
django-redis==3.8.2
django-taggit==0.12.2
django-tinymce==1.5.3
djangorestframework==3.0.3
oauthlib==0.7.2
python-openid==2.2.5
python-social-auth==0.2.1
redis==2.10.3
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
sorl-thumbnail==12.2
wsgiref==0.1.2
django-autoslug==1.7.2
django-ipware==0.1.0